move (future) template to a private header
