+++ include/measurement_kit/dns/dns.hpp
@@ -305,7 +305,7 @@ class Message {
 };
 
 /// Perform a single DNS query
-void query(QueryClass dns_class, QueryType dns_type, std::string name,
+static void query(QueryClass dns_class, QueryType dns_type, std::string name,
            std::function<void(Error, Message)> func,
            Settings settings = {},
            Poller *poller = mk::get_global_poller());
